(search) Reorganize files of datasets report
