+++ applications/search/src/pages/datasets-report-page/publishers-select/publishers-select.scss
