+++ applications/search/src/pages/datasets-report-page/publishers-tree/publishers-tree.component.test.jsx
