+++ applications/search/src/pages/datasets-report-page/report-stats/report-stats.component.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import { ReportStats } from './report-stats.component';
-import aggregateDataset from '../../../../../test/fixtures/aggregateDataset';
+import aggregateDataset from '../../../../test/fixtures/aggregateDataset';
 
 let defaultProps;
 let wrapper;
