+++ applications/search/src/pages/datasets-report-page/report-stats/report-stats.scss
@@ -1,4 +1,4 @@
-@import '../../../../../node_modules/designsystemet/fdk-designsystem-bootstrap3/scss/colors.scss';
+@import '../../../../node_modules/designsystemet/fdk-designsystem-bootstrap3/scss/colors';
 
 .fdk-container-stats {
   background-color: $color0;
