--- applications/search/src/pages/reports-page/reports-page.container.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import { ResultsDatasetsReport } from './results-datasets-report/results-datasets-report.component';
-
-export function ReportsPage() {
-  return (
-    <div>
-      <ResultsDatasetsReport />
-    </div>
-  );
-}
