--- applications/search/src/pages/reports-page/results-datasets-report/results-datasets-report.component.test.jsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import { shallow } from 'enzyme';
-import { ResultsDatasetsReport } from './results-datasets-report.component';
-import aggregateDataset from '../../../../test/fixtures/aggregateDataset';
-
-test('should render ResultsDatasetsReport correctly with props', () => {
-  const wrapper = shallow(<ResultsDatasetsReport />);
-  expect(wrapper).toMatchSnapshot();
-});
-
-test('should render ResultsDatasetsReport correctly with hits', () => {
-  const wrapper = shallow(<ResultsDatasetsReport />);
-  wrapper.setState({
-    aggregateDataset
-  });
-  expect(wrapper).toMatchSnapshot();
-});
